chore(demo-signing): factor out keyManagement.js
