+++ packages/wallet/ui/demo-signing/src/keplr-ledger-demo.js
@@ -1,11 +1,5 @@
 // @ts-check
-import {
-  AminoTypes,
-  SigningStargateClient,
-  defaultRegistryTypes,
-} from '@cosmjs/stargate';
-import { fromBech32, toBech32, fromBase64, toBase64 } from '@cosmjs/encoding';
-import { DirectSecp256k1Wallet, Registry } from '@cosmjs/proto-signing';
+import { SigningStargateClient } from '@cosmjs/stargate';
 import * as tendermintRpcStar from '@cosmjs/tendermint-rpc';
 import { html, render } from 'lit-html';
 import {
@@ -14,17 +8,18 @@ import {
   stakeCurrency,
   stableCurrency,
   bech32Config,
-  SwingsetMsgs,
+  SwingsetMsgs, // TODO: move to keyManagement.js?
 } from './chainInfo.js';
-import { MsgWalletAction } from './gen/swingset/msgs';
+
 import {
-  makeExecMessage,
-  makeFeeGrantMessage,
-  makeGrantWalletActionMessage,
+  makeSigner,
   makeMessagingSigner,
-} from './messagingKey.js';
+  makeExecActionMessages,
+  trivialFee,
+} from './keyManagement.js';
 
 const { freeze } = Object;
+const { Tendermint34Client } = tendermintRpcStar;
 
 const check = {
   /**
@@ -72,56 +67,10 @@ const check = {
   },
 };
 
-/**
- * @typedef {{
- *   typeUrl: '/agoric.swingset.MsgWalletAction',
- *   value: {
- *     owner: string, // base64 of raw bech32 data
- *     action: string,
- *   }
- * }} WalletAction
- */
-
-/** @type {(address: string) => Uint8Array} */
-export function toAccAddress(address) {
-  return fromBech32(address).data;
-}
-
-/** @type {import('@cosmjs/stargate').AminoConverters} */
-const SwingsetConverters = {
-  // '/agoric.swingset.MsgProvision': {
-  //   /* ... */
-  // },
-  [SwingsetMsgs.MsgWalletAction.typeUrl]: {
-    aminoType: SwingsetMsgs.MsgWalletAction.aminoType,
-    toAmino: proto => {
-      const { action, owner } = proto;
-      // NOTE: keep "dictionaries" sorted
-      const amino = {
-        action,
-        owner: toBech32(bech32Config.bech32PrefixAccAddr, fromBase64(owner)),
-      };
-      console.log('@@toAmino:', { proto, amino });
-      return amino;
-    },
-    fromAmino: amino => {
-      const { action, owner } = amino;
-      const proto = { action, owner: toBase64(toAccAddress(owner)) };
-      console.log('@@fromAmino:', { amino, proto });
-      return proto;
-    },
-  },
-};
-
-const aRegistry = new Registry([
-  ...defaultRegistryTypes,
-  [SwingsetMsgs.MsgWalletAction.typeUrl, MsgWalletAction],
-]);
-
 // agoric start local-chain
 const localChainInfo = {
   rpc: 'http://localhost:26657',
-  // rest: api,
+  rest: '@@TODO: fixme',
   chainId: 'agoric',
   chainName: 'Agoric local-chain',
   stakeCurrency,
@@ -136,109 +85,9 @@ const localChainInfo = {
   features: ['stargate', 'ibc-transfer'],
 };
 
-/**
- * @param {ReturnType<typeof makeUI>} ui
- * @param {*} keplr
- * @param {typeof SigningStargateClient.connectWithSigner} connectWithSigner
- */
-const makeSigner = async (ui, keplr, connectWithSigner) => {
-  // const chainId = ui.selectValue('select[name="chainId"]');
-  // console.log({ chainId });
-
-  const chainInfo = localChainInfo;
-  const { chainId } = chainInfo;
-  ui.setValue('*[name="chainId"]', chainId);
-  const { coinMinimalDenom: denom } = stableCurrency;
-
-  await keplr.experimentalSuggestChain(chainInfo);
-  await keplr.enable(chainId);
-
-  // https://docs.keplr.app/api/#get-address-public-key
-  const key = await keplr.getKey(chainId);
-  console.log({ key });
-
-  ui.setChecked('*[name="isNanoLedger"]', key.isNanoLedger);
-
-  // const offlineSigner = await keplr.getOfflineSignerOnlyAmino(chainId);
-  const offlineSigner = await keplr.getOfflineSignerAuto(chainId);
-  console.log({ offlineSigner });
-
-  // Currently, Keplr extension manages only one address/public key pair.
-  const [account] = await offlineSigner.getAccounts();
-  const { address } = account;
-  ui.setValue('*[name="account"]', address);
-
-  const cosmJS = await connectWithSigner(chainInfo.rpc, offlineSigner, {
-    aminoTypes: new AminoTypes(SwingsetConverters),
-    registry: aRegistry,
-  });
-  console.log({ cosmJS });
-
-  const fee = {
-    amount: [{ amount: '100', denom }],
-    gas: '100000', // TODO: estimate gas?
-  };
-  const allowance = '250000'; // 0.25 IST
-
-  return freeze({
-    address, // TODO: address can change
-    authorizeMessagingKey: async (grantee, t0) => {
-      const expiration = t0 / 1000 + 4 * 60 * 60;
-      const msgs = [
-        makeGrantWalletActionMessage(address, grantee, expiration),
-
-        // cosmos support for fee-account in MsgExec hasn't landed yet
-        // https://github.com/cosmos/cosmjs/issues/1155
-        // https://github.com/cosmos/cosmjs/pull/1159
-        // makeFeeGrantMessage(address, grantee, allowance, expiration),
-      ];
-      console.log('sign', { address, msgs, fee });
-      const tx = await cosmJS.signAndBroadcast(address, msgs, fee, '');
-
-      console.log({ tx });
-      return tx;
-    },
-    acceptOffer: async (action, memo) => {
-      const { accountNumber, sequence } = await cosmJS.getSequence(address);
-      console.log({ accountNumber, sequence });
-
-      /** @type {WalletAction} */
-      const act1 = {
-        typeUrl: '/agoric.swingset.MsgWalletAction', // TODO: SpendAction
-        value: {
-          owner: toBase64(toAccAddress(address)),
-          action,
-        },
-      };
-
-      const msgs = [act1];
-
-      // const signDoc = {
-      //   chain_id: chainId,
-      //   account_number: `${accountNumber}`,
-      //   sequence: `${sequence}`,
-      //   fee,
-      //   memo,
-      //   msgs,
-      // };
-
-      // const tx = await cosmJS.signAmino(chainId, account.address, signDoc);
-      // const signerData = { accountNumber, sequence, chainId };
-      console.log('sign', { address, msgs, fee, memo });
-
-      // const tx = await offlineSigner.signAmino(address, signDoc);
-
-      const tx = await cosmJS.signAndBroadcast(address, msgs, fee, memo);
-
-      console.log({ tx });
-      return tx;
-    },
-  });
-};
-
 /** @param {typeof document} document */
 const makeUI = document => {
-  return freeze({
+  const ui = freeze({
     /** @param { string } selector */
     inputValue: selector =>
       check.theInput(document.querySelector(selector)).value,
@@ -264,7 +113,68 @@ const makeUI = document => {
         </ul>`,
         document.querySelector(selector),
       ),
+
+    /**
+     * @param {Awaited<ReturnType<typeof makeSigner>>} s1
+     * @param {Awaited<ReturnType<typeof makeMessagingSigner>>} s2
+     * @param {import('@cosmjs/stargate').SigningStargateClient} lowPrivilegeSigning
+     * @param {tendermintRpcStar.Tendermint34Client} rpcClient
+     */
+    attach: async (s1, s2, rpcClient, lowPrivilegeSigning) => {
+      ui.setValue('*[name="account"]', s1.address);
+      ui.setChecked('*[name="isNanoLedger"]', s1.isNanoLedger);
+
+      ui.onClick('#acceptOffer', async _bev =>
+        s1.acceptOffer(
+          ui.textValue('*[name="spendAction"]'),
+          ui.inputValue('*[name="memo"]'),
+        ),
+      );
+
+      ui.setValue('*[name="messagingAccount"]', s2.address);
+
+      const grants = await s2.queryGrants(
+        {
+          granter: s1.address,
+          msgTypeUrl: SwingsetMsgs.MsgWalletAction.typeUrl,
+        },
+        rpcClient,
+      );
+      console.log('@@@', { grants });
+
+      ui.onClick('#makeMessagingAccount', async _bev =>
+        s1.authorizeMessagingKey(s2.address, Date.now()),
+      );
+
+      ui.onClick('#sendMessages', async _bev => {
+        const { address } = s2;
+        const { accountNumber, sequence } =
+          await lowPrivilegeSigning.getSequence(address);
+        console.log({ accountNumber, sequence });
+
+        const msgs = makeExecActionMessages(
+          s1.address,
+          s2.address,
+          ui.inputValue('*[name="action"]'),
+        );
+        const memo = ui.inputValue('*[name="memo"]');
+
+        const fee = trivialFee();
+
+        console.log('sign', { address, msgs, fee, memo });
+        const tx = await lowPrivilegeSigning.signAndBroadcast(
+          address,
+          msgs,
+          fee,
+          memo,
+        );
+
+        console.log({ tx });
+      });
+    },
   });
+
+  return ui;
 };
 
 window.addEventListener('load', async _ev => {
@@ -276,76 +186,30 @@ window.addEventListener('load', async _ev => {
   const { keplr } = window;
   const ui = makeUI(document);
 
+  const chainInfo = localChainInfo;
+  await keplr.experimentalSuggestChain(chainInfo);
+  const { chainId } = chainInfo;
+  await keplr.enable(chainId);
+  ui.setValue('*[name="chainId"]', chainId);
+
   const s1 = await makeSigner(
-    ui,
+    chainInfo,
     keplr,
     SigningStargateClient.connectWithSigner,
   );
-
-  ui.onClick('#acceptOffer', async _bev =>
-    s1.acceptOffer(
-      ui.textValue('*[name="spendAction"]'),
-      ui.inputValue('*[name="memo"]'),
-    ),
-  );
-
   const s2 = await makeMessagingSigner({ localStorage: window.localStorage });
-  ui.setValue('*[name="messagingAccount"]', s2.address);
 
-  ui.onClick('#makeMessagingAccount', async _bev =>
-    s1.authorizeMessagingKey(s2.address, Date.now()),
-  );
-
-  const chainInfo = localChainInfo;
-  const { Tendermint34Client } = tendermintRpcStar;
   const rpcClient = await Tendermint34Client.connect(chainInfo.rpc);
-  const grants = await s2.queryGrants(
-    { granter: s1.address, msgTypeUrl: SwingsetMsgs.MsgWalletAction.typeUrl },
-    rpcClient,
-  );
-  console.log('@@@', { grants });
 
   const lowPrivilegeClient = await SigningStargateClient.connectWithSigner(
     chainInfo.rpc,
     s2.wallet,
     {
-      registry: aRegistry,
+      registry: s2.registry,
     },
   );
   console.log({ lowPrivilegeClient });
   console.log('low priv signer accounts', await s2.wallet.getAccounts());
 
-  ui.onClick('#sendMessages', async _bev => {
-    const { address } = s2;
-    const { accountNumber, sequence } = await lowPrivilegeClient.getSequence(
-      address,
-    );
-    console.log({ accountNumber, sequence });
-
-    const act1 = {
-      typeUrl: SwingsetMsgs.MsgWalletAction.typeUrl,
-      value: {
-        owner: toBase64(toAccAddress(s1.address)),
-        action: ui.inputValue('*[name="action"]'),
-      },
-    };
-    const msgs = [makeExecMessage(s2.address, [act1], aRegistry)];
-    const memo = ui.inputValue('*[name="memo"]');
-
-    const { coinMinimalDenom: denom } = stableCurrency;
-    const fee = {
-      amount: [{ amount: '0', denom }],
-      gas: '100000', // TODO: estimate gas?
-    };
-
-    console.log('sign', { address, msgs, fee, memo });
-    const tx = await lowPrivilegeClient.signAndBroadcast(
-      address,
-      msgs,
-      fee,
-      memo,
-    );
-
-    console.log({ tx });
-  });
+  ui.attach(s1, s2, rpcClient, lowPrivilegeClient);
 });
